requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
